Releasing v2.2.1 with controllers and testbed
This release has generic closed loop simulation platform to test out
control algorithms with plant models and actuator models.

All linear plant models with an additional possibility of applying some
input conditioning to it that can be used to make it non-linear.

All actuation functions can be wrapped into an Actuator object which is
what this platform uses.

Similarly, all control algorithms can be wrapped as well as long as they
only use the described keyword arguments.

For convinience, childrens of Actuators and Controllers are provided.

A DelayedActuator type is provided that can be used to add a latency
anywhere in the actuator. It is recommended to use this instance to
insert latency for compatibility with provided controllers.

A LinearController is provided that can emulate any linear controller.
See test example to see use for PI controller.

Novel PVLC and MPC algorithms have been provided that help fight against
delayed and non-linear actuator restrictions. See test example for a
quick glance.