Reorganization and introducing Plant as a type
Plant is now an Abstract Type just like Actuators and Controllers. It
will also be callable and should hold any input conditioning inside it.

The InputConditioning is provided as an Abstract Type also and now
system_id and system_id_optimal_inp_cond both return a Plant object
directly.

Note that plant now holds the state of the plant, so it should be used
carefully.

model_evolve will use a deepcopy of passed plant object so that it
doesn't affect the state of the plant.